Struct rman::entries::FileEntry

pub struct FileEntry {

    pub id: i64,
    pub directory_id: i64,
    pub size: u32,
    pub name: String,
    pub tag_bitmask: u64,
    pub unk5: u8,
    pub unk6: u8,
    pub chunk_ids: Vec<i64>,
    pub unk8: u8,
    pub symlink: String,
    pub unk10: u16,
    pub chunking_param_id: u8,
    pub permissions: u8,

}

Single file entry object.

This is identical to the schema in rman-schema and exists to provide a persistent structure for the FileEntry.

Fields
id: i64

Id of the file entry.

directory_id: i64

Id of the directory entry, to which it belongs.

size: u32

Size of the file entry in bytes.

name: String

Name of the file entry.

tag_bitmask: u64

Applicable tags, stored as a bitmask.

unk5: u8

Field with an unknown function and type (it might also be an i8).

unk6: u8

Field with an unknown function and type (it might also be an i8).

chunk_ids: Vec<i64>

A vector of chunk ids that make up the file.

unk8: u8

Field with an unknown function and type (it might also be an i8).

NOTE: seems to always be 1 when a part of .app file on macOS.

symlink: String

Symbolic link of the file entry.

unk10: u16

Field with an unknown function and type (it might also be an i16).

chunking_param_id: u8

Id of the chunking param entry, which provides info about content-defined chunking.

permissions: u8

Permissions for the given file entry.