Enum rman::ManifestError

pub enum ManifestError {
    SeekError(Error),
    InvalidMagicBytes(u32),
    InvalidMajor(u8),
    InvalidMinor(u8),
    InvalidOffset(u32),
    CompressedSizeTooLarge(u32),
    IoError(Error),
    ConversionFailure(TryFromIntError),
    ZstdDecompressError(Error),
    FileParseError(String),
    FlatbufferError(InvalidFlatbuffer),
    ReqwestError(Error),
}

This enum represents all possible errors that may occur when parsing a file.

Variants

SeekError(Error)

The error was caused by a failure to seek to a desired offset.

This error occurs when Seek fails.

InvalidMagicBytes(u32)

The error was caused by invalid magic bytes.

This error occurs when the first four bytes (magic bytes) do not equal 0x52, 0x4D, 0x41 and 0x4E (or R, M, A, N in ascii) respectively.

Usually caused by providing a file that is not a Riot Manifest file.

InvalidMajor(u8)

Available on crate feature version_error only.

The error was caused by invalid major version.

This error occurs when the major version in the file header doesnโt equal 2.

Should only occur if the manifest format gets a major change or an update, Parser may no longer function if this happens.

NOTE: The feature version_error must be enabled for this error to occur.

InvalidMinor(u8)

Available on crate feature version_error only.

The error was caused by invalid minor version.

This error occurs when the minor version in the file header doesnโt equal 0.

Should only occur if the manifest format gets a minor change or an update. Parser should still be functional if this happens,

NOTE: The feature version_error must be enabled for this error to occur.

InvalidOffset(u32)

The error was caused by an invalid offset.

This error occurs when the offset is smaller or larger than the file itself.

Should never happen for official, Riot-made manifests.

CompressedSizeTooLarge(u32)

The error was caused by compressed size being too large.

This error occurs when the compressed size is larger than the file itself.

Should never happen for official, Riot-made manifests.

IoError(Error)

The error was caused by a failure to read or write bytes on an IO stream.

This error occurs when read_exact, any read_ method in [byteorder::ReadBytesExt], or File::open fails.

Usually caused by invalid or inaccessible path or unexpected eof when parsing a header.

ConversionFailure(TryFromIntError)

The error was caused by a failure to convert from one number type to another.

This error occurs when the conversion from or into usize fails.

Should never fail on 32-bit (or higher) targets.

ZstdDecompressError(Error)

The error was caused by a failure to decompress zstd data.

This error occurs when [decompress][zstd::bulk::decompress] fails.

Should never happen for official, Riot-made manifests.

FileParseError(String)

The error was caused by a failure to parse FileEntry into File.

This error occurs when either a directory_id or parent_id points to an invalid DirectoryEntry or when chunk_id refers to an invalid ChunkEntry.

Should never happen for official, Riot-made manifests.

FlatbufferError(InvalidFlatbuffer)

The error was caused by an invalid flatbuffer.

This error occurs when a flatbuffer fails to verify.

Should never happen for official, Riot-made manifests.

ReqwestError(Error)

The error was caused by a failure to process a Request.

This error occurs when Client::send() fails.

Usually occurs when there is no internet connection, or when an invalid bundle url was provided.